Turn breadcrumb molecules into exported components
