+++ packages/designmanual/stories/molecules/breadcrumbs.tsx
@@ -7,12 +7,8 @@
  */
 
 import React from 'react';
-import { Breadcrumb, IndexedBreadcrumbItem } from '@ndla/ui';
-import styled from '@emotion/styled';
-import { breakpoints, colors, fonts, mq, spacing } from '@ndla/core';
-import { ChevronRight } from '@ndla/icons/src/common';
-import { Home } from '@ndla/icons/lib/common';
-import SafeLink from '@ndla/safelink';
+import { Breadcrumb } from '@ndla/ui';
+import { HeaderBreadcrumb, HomeBreadcrumb } from '@ndla/ui/src';
 
 const items = [
   {
@@ -38,156 +34,19 @@ interface BreadcrumbDefaultProps {
 }
 
 export const BreadcrumbDefault = ({ autoCollapse }: BreadcrumbDefaultProps) => (
-  <Breadcrumb autoCollapse={autoCollapse} items={[...items]} />
+  <Breadcrumb autoCollapse={autoCollapse} items={items} />
 );
 
-const StyledHeaderSafeLink = styled(SafeLink)`
-  ${fonts.sizes(14)};
-  font-weight: ${fonts.weight.bold};
-  color: ${colors.brand.primary};
-`;
-
-const StyledIconSafeLink = styled(SafeLink)`
-  box-shadow: none;
-  border-bottom: none;
-`;
-const StyledBlueRightChevron = styled(ChevronRight)`
-  color: ${colors.brand.primary};
-  margin: ${spacing.xxsmall};
-`;
-const StyledBlueSpan = styled.span`
-  color: ${colors.brand.primary};
-`;
-const StyledBlueSafeLink = styled(SafeLink)`
-  color: ${colors.brand.primary};
-`;
-
 export const BreadcrumbWithHeader = () => {
-  const renderItem = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return <StyledBlueSpan>{item.name}</StyledBlueSpan>;
-    }
-    return <StyledBlueSafeLink to={item.to}>{item.name}</StyledBlueSafeLink>;
-  };
-
-  const renderSeparator = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return null;
-    }
-    return <StyledBlueRightChevron />;
-  };
-
-  return (
-    <>
-      <StyledHeaderSafeLink to={items[0].to}>{items[0].name}</StyledHeaderSafeLink>
-      <Breadcrumb items={items.slice(1)} renderItem={renderItem} renderSeparator={renderSeparator} autoCollapse />
-    </>
-  );
+  return <HeaderBreadcrumb items={items} />;
 };
-
-const StyledSeparator = styled.div`
-  ${fonts.sizes('14px')};
-  margin: 0 ${spacing.small};
-  user-select: none;
-  ${mq.range({ until: breakpoints.tablet })} {
-    display: none;
-  }
-`;
-const StyledHome = styled(Home)`
-  width: 20px;
-  height: 20px;
-  color: ${colors.text.primary};
-`;
-const StyledDarkRightChevron = styled(ChevronRight)`
-  color: ${colors.text.primary};
-  margin: ${spacing.xxsmall};
-`;
-const StyledDarkSpan = styled.span`
-  color: ${colors.text.primary};
-`;
-const StyledDarkSafeLink = styled(SafeLink)`
-  color: ${colors.text.primary};
-`;
 
 export const BreadcrumbWithHome = () => {
-  const renderItem = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return <StyledDarkSpan>{item.name}</StyledDarkSpan>;
-    }
-    if (item.index === 0) {
-      return (
-        <StyledIconSafeLink to={item.to}>
-          <StyledHome />
-        </StyledIconSafeLink>
-      );
-    }
-    return <StyledDarkSafeLink to={item.to}>{item.name}</StyledDarkSafeLink>;
-  };
-
-  const renderSeparator = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return null;
-    }
-    if (item.index === 0) {
-      return <StyledSeparator>|</StyledSeparator>;
-    }
-    return <StyledDarkRightChevron />;
-  };
-
-  return <Breadcrumb items={items} renderItem={renderItem} renderSeparator={renderSeparator} />;
+  return <HomeBreadcrumb items={items} />;
 };
 
-const StyledWhiteSeparator = styled.div`
-  ${fonts.sizes('14px')};
-  margin: 0 ${spacing.small};
-  user-select: none;
-  color: ${colors.white};
-  ${mq.range({ until: breakpoints.tablet })} {
-    display: none;
-  }
-`;
-const StyledWhiteHome = styled(Home)`
-  width: 20px;
-  height: 20px;
-  color: ${colors.white};
-`;
-const StyledWhiteRightChevron = styled(ChevronRight)`
-  color: ${colors.white};
-  margin: ${spacing.xxsmall};
-`;
-const StyledWhiteSpan = styled.span`
-  color: ${colors.white};
-`;
-const StyledWhiteSafeLink = styled(SafeLink)`
-  color: ${colors.white};
-`;
-
 export const BreadcrumbWhiteWithHome = () => {
-  const renderItem = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return <StyledWhiteSpan>{item.name}</StyledWhiteSpan>;
-    }
-    if (item.index === 0) {
-      return (
-        <StyledIconSafeLink to={item.to}>
-          <StyledWhiteHome />
-        </StyledIconSafeLink>
-      );
-    }
-    return <StyledWhiteSafeLink to={item.to}>{item.name}</StyledWhiteSafeLink>;
-  };
-
-  const renderSeparator = (item: IndexedBreadcrumbItem, totalCount: number) => {
-    if (item.index === totalCount - 1) {
-      return null;
-    }
-    if (item.index === 0) {
-      return <StyledWhiteSeparator>|</StyledWhiteSeparator>;
-    }
-    return <StyledWhiteRightChevron />;
-  };
-
-  return <Breadcrumb items={items} renderItem={renderItem} renderSeparator={renderSeparator} />;
+  return <HomeBreadcrumb items={items} light />;
 };
 
 export default BreadcrumbDefault;
