+++ packages/ndla-ui/src/Breadcrumb/index.ts
@@ -10,4 +10,8 @@ import Breadcrumb from './Breadcrumb';
 
 export type { SimpleBreadcrumbItem, IndexedBreadcrumbItem } from './BreadcrumbItem';
 
+export { default as HeaderBreadcrumb } from './HeaderBreadcrumb';
+
+export { default as HomeBreadcrumb } from './HomeBreadcrumb';
+
 export default Breadcrumb;
